Captura do Windows: respeita a tela escolhida e identifica janelas
Duas coisas quebradas no `windows.rs`, as duas silenciosas:

`start()` ignorava `config.source` e ligava sempre em `Monitor::primary()`. Escolher
uma janela no seletor e transmitir o monitor inteiro nao e um detalhe de UI: e mandar
e-mail e conversa para a sala quando a pessoa pediu para mandar so o jogo.

`windows()` devolvia `id: 0` para todas as janelas. Sem identidade, escolher a segunda
da lista transmitia a primeira. Agora o id e o HWND — e por isso `Window.id` e
`CaptureSource::Window` passaram de u32 para u64, ja que HWND e ponteiro. No macOS o
CGWindowID de 32 bits so e alargado.

Verificado com `cargo check --target x86_64-pc-windows-msvc` a partir do macOS: o
`capture` compila para Windows sem SDK porque nao tem dependencia em C. Workspace e
testes seguem verdes no macOS.

+++ native/crates/capture/src/macos.rs
@@ -128,7 +128,7 @@ impl MacCapturer {
             .into_iter()
             .filter(|window| window.title().is_some_and(|title| !title.is_empty()))
             .map(|window| Window {
-                id: window.window_id(),
+                id: u64::from(window.window_id()),
                 title: window.title().unwrap_or_default(),
                 application: window
                     .owning_application()
@@ -158,7 +158,7 @@ impl MacCapturer {
                 let window = content
                     .windows()
                     .into_iter()
-                    .find(|window| window.window_id() == id)
+                    .find(|window| u64::from(window.window_id()) == id)
                     .ok_or(CaptureError::NoDisplay)?;
 
                 SCContentFilter::create().with_window(&window).build()
@@ -219,7 +219,7 @@ impl MacCapturer {
                 let window = content
                     .windows()
                     .into_iter()
-                    .find(|window| window.window_id() == id)
+                    .find(|window| u64::from(window.window_id()) == id)
                     .ok_or(CaptureError::NoDisplay)?;
 
                 SCContentFilter::create().with_window(&window).build()
